requirements.txt
# ── Vita AI – Python Dependencies ──────────────────────────────
# Requires Python 3.11.x  (tested with 3.11.9)
# Install with:  pip install -r requirements.txt

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
websockets>=12.0          # required for WebSocket /ws/face-scan endpoint

# Data / schema
pydantic>=2.0

# Database (SQLAlchemy + SQLite)
sqlalchemy>=2.0

# Authentication – Code contributed by Manogna (JWT + bcrypt)
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
bcrypt>=4.0,<5.0

# Encryption – Security/encryption logic adapted from Manogna's backend
cryptography>=41.0

# Numerical
numpy>=1.24,<2.0

# Computer vision
opencv-python>=4.8
mediapipe>=0.10

# Audio processing
librosa>=0.10
soundfile>=0.12
resampy>=0.4

# NLP / Transformers
transformers>=4.30
torch>=2.0
sentencepiece>=0.1.99   # required for BioBERT tokenizer

# ML utilities (reserved for future score-fusion model)
scikit-learn>=1.3

# Optional: XGBoost for ML-based score fusion
xgboost>=2.0

# TensorFlow + TF-Hub for YAMNet audio classifier
tensorflow>=2.13
tensorflow-hub>=0.14

# Signal processing
scipy>=1.11

# Primary rPPG backend (Open-rPPG)
open-rppg>=0.1.1

# Evaluation harness (optional but recommended)
pandas>=2.0
matplotlib>=3.7
seaborn>=0.13

# HTTP client
requests>=2.31

# Testing
pytest>=7.4
httpx>=0.24            # for FastAPI TestClient
